etl/requirements.txt
async-timeout==4.0.2
certifi==2022.12.7
elastic-transport==8.4.0
elasticsearch==8.6.2
flake8==6.0.0
flake8-import-order==0.18.2
isort==5.12.0
mccabe==0.7.0
psycopg2-binary==2.9.5
pycodestyle==2.10.0
pydantic==1.10.5
pyflakes==3.0.1
python-dotenv==0.21.1
redis==4.5.1
typing_extensions==4.5.0
urllib3==1.26.14
